requirements.txt
Flask
gunicorn
psycopg2-binary
pydantic
pytest
python-dotenv
PyYAML
